have BoxCluster keep track of a (non-static) ID for the coordinate pair set use a long for the ID instead of an int allow ClusterSum (and a few friends) to explicitly disable caching the values dump ClusterCoupled dump revertF(), dirtyAtom

+++ etomica.apps/etomica/virial/ClusterChainHS.java
@@ -17,7 +17,7 @@ public class ClusterChainHS implements ClusterAbstract {
     protected final long[][] nC;
     protected final double[][] fValues;
     protected double ringFrac, chainFrac;
-    protected int cPairID = -1, lastCPairID = -1;
+    protected long cPairID = -1, lastCPairID = -1;
     protected double value, lastValue;
     public final boolean old = true;
     
@@ -63,7 +63,7 @@ public class ClusterChainHS implements ClusterAbstract {
 
     public double value(BoxCluster box) {
         CoordinatePairSet cPairs = box.getCPairSet();
-        int thisCPairID = cPairs.getID();
+        long thisCPairID = cPairs.getID();
         if (thisCPairID == cPairID) {
             return value;
         }
